+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fix fill booking_id when chaing status 
+
 ### 1.1.1: April 5th, 2021
 * Add barcode field to course sql table
 * Barcode scan JSOn returns error messages