+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 1.1.2: April 6th, 2021
 * Fix fill booking_id when changing status
 * Fix sending JSON data with default function, not error or success  
 * Remove Gravity Forms overflow hidden